Move RegistrationsDisabled scene into Error subdirectory
This is done to re-use code for other errors later on.
