+++ apps/client/src/scenes/Error/RegistrationsDisabled/index.ts
